Back-End/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
psycopg2
python-dotenv
supabase
Pillow
